Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.26.0
motor==3.5.1
pymongo==4.8.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.6
websockets==12.0